+++ neurodamus/core/_mpi.py
@@ -35,11 +35,8 @@ class MPI(object):
         # When using MPI (and more than 1 rank) we need to MPIAbort on exception to avoid deadlocks
         def excepthook(etype, value, tb):
             time.sleep(0.1 * cls._rank)  # Order errors
-            logging.critical(str(value))
+            logging.critical(str(value), exc_info=True)
             pc.allreduce(1, 1)  # Share error state
-            # Print and cleanup exception
-            sys.__excepthook__(etype, value, tb)
-            Neuron.quit()  # With 'special' terminates right here
             sys.exit(1)
         sys.excepthook = excepthook
 
@@ -49,8 +46,8 @@ class MPI(object):
         res = cls._pc.allreduce(0, 1)
         if res > 0:
             if MPI.rank == 0:
-                logging.critical("Another rank raised an irrecoverable error. Check log file.")
-            Neuron.quit()
+                logging.critical("Another rank raised an irrecoverable error. "
+                                 "Check log file. Terminating.")
             sys.exit(1)
 
     @property